web/back/pyproject.toml
[tool.poetry]
name = "web"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
boto3 = "^1.28.76"
fastapi = {extras = ["all"], version = "^0.104.1"}
ydb = "*"
python-telegram-bot = "*"
yarl = "^1.9.4"
cachetools = "^5.3.2"
pyjwt = "^2.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
isort = "^5.12.0"
black = "^23.10.1"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
